Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/events/invalid-uncompiled-raw-handler-compiled-late-expected.txt

CONSOLE MESSAGE: line 24: SyntaxError: Parser error
CONSOLE MESSAGE: line 21: SyntaxError: Parser error

FAIL Invalid uncompiled raw handlers should only be compiled when about to call them. assert_array_equals: lengths differ, expected 3 got 4
